# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank system
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /blocks/process:
    put:
      tags:
      - Card Block Management
      summary: Process a card block request
      description: Allows an administrator to approve or reject a block request
      operationId: processBlockRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessBlockRequestDto"
        required: true
      responses:
        "200":
          description: Block request processed successfully
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Validation failed
                code: 400
                validationErrors:
                - field: status
                  message: Status cannot be null
        "404":
          description: Block request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Request to block by id: 1 not found"
                code: 404
...
  /payments:
    post:
      tags:
      - Payment Management
      summary: Transfer money between cards
      description: Allows a user to transfer money between their own cards
      operationId: transferMoney
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentDto"
        required: true
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "400":
          description: Invalid input data or insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                Validation Error:
                  description: Validation Error
                  value:
                    timestamp: 2024-07-30T12:34:56.789Z
                    message: Validation failed
                    code: 400
                    validationErrors:
                    - field: amount
                      message: Amount must not be less than zero
                Insufficient Funds:
                  description: Insufficient Funds
                  value:
                    timestamp: 2024-07-30T12:34:56.789Z
                    message: There are not enough funds on the card
                    code: 400
        "403":
          description: Card is blocked or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 is blocked for operation"
                code: 403
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
  /blocks:
    post:
      tags:
      - Card Block Management
      summary: Create a card block request
      description: Allows a user to request blocking their own card
      operationId: createBlockRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBlockRequestDto"
        required: true
      responses:
        "201":
          description: Block request created successfully
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: Validation failed
                code: 400
                validationErrors:
                - field: reason
                  message: Reason should not be empty
        "403":
          description: Card is already blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: You cannot block an inactive card
                code: 403
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
...
  /cards:
    get:
      tags:
      - User Card Management
      summary: Get all user cards
      description: Retrieves a paginated list of all cards belonging to the authenticated
        user
      operationId: getUserCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "balance,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
  /cards/{id}:
    get:
      tags:
      - User Card Management
      summary: Get card by ID
      description: "Retrieves a specific card by its ID, if it belongs to the authenticated\
        \ user"
      operationId: getCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
  /blocks/{id}:
    get:
      tags:
      - Card Block Management
      summary: Find block request by ID
      description: Retrieves a specific block request by its ID
      operationId: findCardBlockRequestById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found block request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardBlockRequestDto"
        "404":
          description: Block request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Request for blocking by id: 1 not found"
                code: 404
  /blocks/processed-by/{adminId}:
    get:
      tags:
      - Card Block Management
      summary: Find block requests processed by an admin
      description: Retrieves a paginated list of block requests processed by a specific
        administrator
      operationId: findCardBlockRequestByProcessed
      parameters:
      - name: adminId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "createAt,ASC"
          items:
            type: string
      responses:
        "200":
          description: Found block requests
          content:
            application/json: {}
...
  /admin/cards/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get card by ID
      description: Retrieves card details by its ID
      operationId: getCardById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
    delete:
      tags:
      - Admin Card Management
      summary: Delete card by ID
      description: Deletes a card by its ID
      operationId: deleteCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Card deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
  /admin/cards/user/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get all cards for a user
      description: Retrieves a paginated list of all cards for a specified user
      operationId: getUserCards_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "validityPeriod,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
...
  /users/{id}:
    delete:
      tags:
      - User Management
      summary: Delete user by ID
      description: Deletes a user by their ID
      operationId: deleteUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: User deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "User by id: 1 not found"
                code: 404
components:
  schemas:
    ProcessBlockRequestDto:
      type: object
      description: Request to process a card block
      properties:
        requestId:
          type: integer
          format: int64
          description: ID of the block request to process
          example: 1
        status:
          type: string
          description: New status for the block request
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        adminId:
          type: integer
          format: int64
          description: ID of the admin processing the request
          example: 1
      required:
      - adminId
      - requestId
      - status
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
        message:
          type: string
          description: Error message
          example: Validation failed
        code:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        validationErrors:
          type: array
          description: List of validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      description: Details of a validation error
      properties:
        field:
          type: string
          description: Field where the error occurred
          example: username
        message:
          type: string
          description: Error message for the field
          example: Username must not be empty
...
    PaymentDto:
      type: object
      description: Request for a payment transfer
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID of the card to transfer from
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID of the card to transfer to
          example: 2
        amount:
          type: number
          description: Amount to transfer
          example: 100.5
        comment:
          type: string
          description: Comment for the payment
          example: Payment for goods
      required:
      - fromCardId
      - toCardId
    CardDto:
      type: object
      description: Card details
      properties:
        cardId:
          type: integer
          format: int64
          description: Card ID
          example: 1
        cardNumber:
          type: string
          description: Masked card number
          example: '**** **** **** 9010'
        validityPeriod:
          type: string
          format: date
          description: Expiration date of the card
          example: 2028-12-31
        statusCard:
          type: string
          description: Current status of the card
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - PENDING_BLOCKED
        balance:
          type: number
          description: Current balance of the card
          example: 1000.0
    CreateCardBlockRequestDto:
      type: object
      description: Request to create a card block
      properties:
        cardId:
          type: integer
          format: int64
          description: ID of the card to be blocked
          example: 1
        reason:
          type: string
          description: Reason for blocking the card
          example: I lost my card
          maxLength: 255
          minLength: 5
      required:
      - reason
...
    UserDto:
      type: object
      description: User details
      properties:
        id:
          type: integer
          format: int64
          description: User ID
          example: 1
        username:
          type: string
          description: Username
          example: john_doe
        role:
          type: string
          description: User role
          enum:
          - ROLE_USER
          - ROLE_ADMIN
        cards:
          type: array
          description: List of user's cards
          items:
            $ref: "#/components/schemas/CardDto"
        accountEnable:
          type: boolean
          description: Is the user account enabled
          example: true
        accountLocked:
          type: boolean
          description: Is the user account locked
          example: false
...
    CardBlockRequestDto:
      type: object
      description: Card block request details
      properties:
        id:
          type: integer
          format: int64
          description: Request ID
          example: 1
        card:
          $ref: "#/components/schemas/CardDto"
          description: Card information
        requester:
          $ref: "#/components/schemas/UserDto"
          description: User who requested the block
        reason:
          type: string
          description: Reason for blocking
          example: Card lost
        processedBy:
          $ref: "#/components/schemas/UserDto"
          description: Administrator who processed the request
        processedAt:
          type: string
          format: date-time
          description: Timestamp of when the request was processed
        blockRequestStatus:
          type: string
          description: Status of the block request
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        createAt:
          type: string
          format: date-time
          description: Timestamp of when the request was created
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT